# Matrix inversion speedup