From https://github.com/PX4/PX4-Matrix/pull/165

Also fixes the math benchmark: trying to measure a single instruction with perf counters does not work (similar for atomic ops).

The results make more sense now when measuring over a number of ops (e.g. float div slower than add/sub/mul, double is at least an order of magnitude slower than float).
```
INFO  [microbench] RUNNING TEST: time_single_precision_float
float add (10k ops): 10 events, 4887us elapsed, 488.70us avg, min 488us max 489us 0.483us rms
float sub (10k ops): 10 events, 4890us elapsed, 489.00us avg, min 489us max 489us 0.005us rms
float mul (10k ops): 10 events, 4888us elapsed, 488.80us avg, min 488us max 489us 0.421us rms
float div (10k ops): 10 events, 12625us elapsed, 1262.50us avg, min 1262us max 1263us 0.532us rms
float sqrt (1k ops): 10 events, 1202us elapsed, 120.20us avg, min 120us max 121us 0.421us rms
INFO  [microbench] TEST PASSED: time_single_precision_float

INFO  [microbench] RUNNING TEST: time_single_precision_float_trig
sinf() (1k ops): 10 events, 10800us elapsed, 1080.00us avg, min 1080us max 1080us 0.068us rms
cosf() (1k ops): 10 events, 10500us elapsed, 1050.00us avg, min 1050us max 1050us 0.063us rms
tanf() (1k ops): 10 events, 19012us elapsed, 1901.20us avg, min 1901us max 1902us 0.410us rms
acosf() (1k ops): 10 events, 4901us elapsed, 490.10us avg, min 490us max 491us 0.315us rms
asinf() (1k ops): 10 events, 4605us elapsed, 460.50us avg, min 460us max 461us 0.526us rms
atan2f() (1k ops): 10 events, 13537us elapsed, 1353.70us avg, min 1353us max 1354us 0.491us rms
INFO  [microbench] TEST PASSED: time_single_precision_float_trig

INFO  [microbench] RUNNING TEST: time_double_precision_float
double add (1k ops): 10 events, 5863us elapsed, 586.30us avg, min 586us max 587us 0.483us rms
double sub (1k ops): 10 events, 5743us elapsed, 574.30us avg, min 574us max 575us 0.482us rms
double mul (1k ops): 10 events, 4589us elapsed, 458.90us avg, min 458us max 459us 0.315us rms
double div (100 ops): 10 events, 3780us elapsed, 378.00us avg, min 378us max 378us 0.009us rms
double sqrt (100 ops): 10 events, 1180us elapsed, 118.00us avg, min 118us max 118us 0.006us rms
INFO  [microbench] TEST PASSED: time_double_precision_float

INFO  [microbench] RUNNING TEST: time_double_precision_float_trig
sin() (100 ops): 10 events, 15960us elapsed, 1596.00us avg, min 1596us max 1596us   nanus rms
cos() (100 ops): 10 events, 17345us elapsed, 1734.50us avg, min 1734us max 1735us 0.517us rms
tan() (100 ops): 10 events, 23179us elapsed, 2317.90us avg, min 2317us max 2318us 0.331us rms
acos() (100 ops): 10 events, 1507us elapsed, 150.70us avg, min 150us max 151us 0.483us rms
asin() (100 ops): 10 events, 1726us elapsed, 172.60us avg, min 172us max 173us 0.516us rms
atan2() (100 ops): 10 events, 27132us elapsed, 2713.20us avg, min 2713us max 2714us 0.441us rms
sqrt() (100 ops): 10 events, 1182us elapsed, 118.20us avg, min 118us max 119us 0.421us rms
INFO  [microbench] TEST PASSED: time_double_precision_float_trig

INFO  [microbench] RUNNING TEST: time_8bit_integers
int8 add (10k ops): 10 events, 4948us elapsed, 494.80us avg, min 494us max 495us 0.422us rms
int8 sub (10k ops): 10 events, 4944us elapsed, 494.40us avg, min 494us max 495us 0.517us rms
int8 mul (10k ops): 10 events, 4945us elapsed, 494.50us avg, min 494us max 495us 0.527us rms
int8 div (10k ops): 10 events, 4888us elapsed, 488.80us avg, min 488us max 489us 0.421us rms
INFO  [microbench] TEST PASSED: time_8bit_integers

INFO  [microbench] RUNNING TEST: time_16bit_integers
int16 add (10k ops): 10 events, 4885us elapsed, 488.50us avg, min 488us max 489us 0.527us rms
int16 sub (10k ops): 10 events, 4888us elapsed, 488.80us avg, min 488us max 489us 0.421us rms
int16 mul (10k ops): 10 events, 4889us elapsed, 488.90us avg, min 488us max 489us 0.316us rms
int16 div (10k ops): 10 events, 7266us elapsed, 726.60us avg, min 726us max 727us 0.515us rms
INFO  [microbench] TEST PASSED: time_16bit_integers

INFO  [microbench] RUNNING TEST: time_32bit_integers
int32 add (10k ops): 10 events, 4290us elapsed, 429.00us avg, min 429us max 429us   nanus rms
int32 sub (10k ops): 10 events, 4291us elapsed, 429.10us avg, min 429us max 430us 0.315us rms
int32 mul (10k ops): 10 events, 4292us elapsed, 429.20us avg, min 429us max 430us 0.421us rms
int32 div (10k ops): 10 events, 4888us elapsed, 488.80us avg, min 488us max 489us 0.422us rms
INFO  [microbench] TEST PASSED: time_32bit_integers

INFO  [microbench] RUNNING TEST: time_64bit_integers
int64 add (1k ops): 10 events, 1199us elapsed, 119.90us avg, min 119us max 120us 0.316us rms
int64 sub (1k ops): 10 events, 1273us elapsed, 127.30us avg, min 127us max 128us 0.482us rms
int64 mul (1k ops): 10 events, 1008us elapsed, 100.80us avg, min 100us max 101us 0.421us rms
int64 div (1k ops): 10 events, 6864us elapsed, 686.40us avg, min 686us max 687us 0.516us rms
INFO  [microbench] TEST PASSED: time_64bit_integers
```
However there's a remaining issue: loading and storing from/to memory uses additional runtime, distorting the results (this could be fixed by using inline assembly). So I'm inclined to just remove the tests to avoid incorrect interpretation.

Out of interest I also checked cos/sin from cmsis:
```
sinf() (1k ops): 10 events, 8714us elapsed, 871.40us avg, min 871us max 872us 0.515us rms
cosf() (1k ops): 10 events, 8357us elapsed, 835.70us avg, min 835us max 836us 0.480us rms
arm_sin() (1k ops): 10 events, 2933us elapsed, 293.30us avg, min 293us max 294us 0.483us rms
arm_cos() (1k ops): 10 events, 3110us elapsed, 311.00us avg, min 311us max 311us   nanus rms
```
So we would gain something there, but the result is also less accurate.