@dagar 2021-09-03
Thanks for fixing the microbenchmarks for the quick operations. This was only cobbled together as a crude way to casually compare things between boards and be aware of the more expensive operations (eg atan2) as double precision hardware became more common. That’s also why the “tests” don’t actually have failure conditions.

If you’re worried about misinterpretation you could add a warning about the timing information. 


> Out of interest I also checked cos/sin from cmsis:

I actually tried this as well, but I wasn't sure if it was worth the hassle (platform differences, how it slips into the build, cmsis dependency in general that I'd still be happy to purge). Maybe we should take a quick look to estimate the potential value in aggregate?