@dagar 2021-09-05
Finally fitting everywhere.

The last failure is the test rack where `sensors` reports a timeout when the microbench (with longer critical sections) runs. 

<img width="786" alt="Screen Shot 2021-09-05 at 6 16 32 PM" src="https://user-images.githubusercontent.com/84712/132142805-e29c1126-b61d-4f96-ba9e-7dfcc4a84118.png">

I'll stop the module first.