@bkueng 2021-09-07
Thanks.

> Maybe we should take a quick look to estimate the potential value in aggregate?

I found the difference large enough to further explore it and then decide if it's worth the hassle. But it might overall not make a significant difference.